refactor and show grace period
